ejercicio_1/requirements.txt
fastapi
pydantic
python-dotenv
SQLAlchemy
uvicorn
pymysql
cryptography